# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface DriveTrack {
  id: string;
  title: string;
  /**
     * @minimum 1
     * @maximum 25165824
     */
  size: number;
}

export interface PlaylistSummary {
  id: string;
  name: string;
}

export interface SavePlaylistInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minItems 1
     * @maxItems 20
     */
  trackIds: string[];
}

export interface SavedPlaylist {
  id: string;
  name: string;
  /**
     * @minItems 1
     * @maxItems 20
     */
  tracks: DriveTrack[];
}

export interface OwnerLoginInput {
  /** @maxLength 256 */
  password: string;
}

export interface OwnerStatus {
  owner: boolean;
  configured: boolean;
}