# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
servers:
  - url: /api
paths:
  /healthz:
    get:
      operationId: healthCheck
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /playlists/library:
    get:
      operationId: getMusicLibrary
      responses:
        "200":
          description: Library tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriveTrack"
  /playlists/audio/{id}:
    get:
      operationId: getDriveAudio
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: MP3, at most 24 MB and 12 minutes
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
  /playlists:
    get:
      operationId: listSharedPlaylists
      responses:
        "200":
          description: Shared playlist names
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistSummary"
    post:
      operationId: saveSharedPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePlaylistInput"
      responses:
        "201":
          description: Saved without overwriting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistSummary"
  /playlists/{id}:
    get:
      operationId: loadSharedPlaylist
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Ordered referenced tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPlaylist"
  /playlists/owner:
    get:
      operationId: getOwnerStatus
      responses:
        "200":
          description: Owner session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerStatus"
    post:
      operationId: ownerSignIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerLoginInput"
      responses:
        "200":
          description: Owner session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerStatus"
    delete:
      operationId: ownerSignOut
      responses:
        "204":
          description: Signed out
  /playlists/upload:
    post:
      operationId: uploadOwnerMp3
      parameters:
        - name: X-Filename
          in: header
          required: true
          schema:
            type: string
            description: URI-encoded MP3 filename
      requestBody:
        required: true
        content:
          audio/mpeg:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Uploaded to the designated music folder; owner cookie required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriveTrack"
components:
  parameters:
    FileId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-Za-z0-9_-]{1,200}$"
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    DriveTrack:
      type: object
      required: [id, title, size]
      properties:
        id:
          type: string
        title:
          type: string
        size:
          type: integer
          minimum: 1
          maximum: 25165824
    PlaylistSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
    SavePlaylistInput:
      type: object
      additionalProperties: false
      required: [name, trackIds]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        trackIds:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
    SavedPlaylist:
      type: object
      required: [id, name, tracks]
      properties:
        id:
          type: string
        name:
          type: string
        tracks:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/DriveTrack"
    OwnerLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          maxLength: 256
    OwnerStatus:
      type: object
      required: [owner, configured]
      properties:
        owner:
          type: boolean
        configured:
          type: boolean