# lib/api-zod/src/generated/types/driveTrack.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.2.0
 */

export interface DriveTrack {
  id: string;
  title: string;
  /**
     * @minimum 1
     * @maximum 25165824
     */
  size: number;
}