# lib/api-zod/src/generated/types/savePlaylistInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.2.0
 */

export interface SavePlaylistInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minItems 1
     * @maxItems 20
     */
  trackIds: string[];
}