# lib/api-zod/src/generated/types/savedPlaylist.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.2.0
 */
import type { DriveTrack } from './driveTrack';

export interface SavedPlaylist {
  id: string;
  name: string;
  /**
     * @minItems 1
     * @maxItems 20
     */
  tracks: DriveTrack[];
}